new try

+++ hashmap.c
@@ -24,7 +24,7 @@ Pair * createPair( char * key,  void * value) {
 }
 
 long hash( char * key, long capacity) {
-    unsigned long hash = 0;
+    size_t hash = 0;
      char * ptr;
     for (ptr = key; *ptr != '\0'; ptr++) {
         hash += hash*32 + tolower(*ptr);
